modification des messages d'erreurs

+++ controllers/sauces.js
@@ -9,7 +9,7 @@ exports.createSauce = (req, res, next) => {
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
     sauce.save()
-        .then(() => res.status(201).json({ message: 'Objet enregistré !'}))
+        .then(() => res.status(201).json({ message: 'Enregistrement effectué !'}))
         .catch(error => res.status(400).json({ error }));
     };
 
@@ -26,25 +26,37 @@ exports.getAllSauces = (req, res) => {
   };
 
 exports.modifySauce = (req, res, next) => {
-    const sauceObject = req.file ?
-    {
-        ...JSON.parse(req.body.sauce),
-        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
-    } : { ...req.body };
-    Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Objet modifié !'}))
-        .catch(error => res.status(400).json({ error }));
+    Sauce.findOne({ _id: req.params.id })
+    .then(sauce => {
+        const sauceObject = req.file ?
+        {
+            ...JSON.parse(req.body.sauce),
+            imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+        } : { ...req.body };
+        if (sauce.userId === req.token.userId) {
+            Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                .then(() => res.status(200).json({ message: 'Modification éffectuée !'}))
+                .catch(error => res.status(400).json({ error }));
+        /*} else {
+            res.status(403).json({ message: 'Modification non autorisée !'})
+        */}
+    })
+    .catch(error => res.status(500).json({ error }));
 };
 
 exports.deleteSauce = (req, res, next) => {     
     Sauce.findOne({ _id: req.params.id })
     .then(sauce => {
-        const filename = sauce.imageUrl.split('/images/')[1];
-        fs.unlink(`images/${filename}`, () => {
-            Sauce.deleteOne({ _id: req.params.id })
-                .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-                .catch(error => res.status(400).json({ error }));
-      });
+        if (sauce.userId === req.token.userId) {
+            const filename = sauce.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+                Sauce.deleteOne({ _id: req.params.id })
+                    .then(() => res.status(200).json({ message: ' Suppression effectuée !'}))
+                    .catch(error => res.status(400).json({ error }));
+            });
+        } else {
+            res.status(401).json({ message: 'Suppression non autorisée !'})
+        }  
     })
     .catch(error => res.status(500).json({ error }));
 };
